Cleaning up code
Splitting stray functions into a separate location.
