+++ plugins/CIBModule.cpp
@@ -32,96 +32,25 @@
 #include <cstdint>
 #include <string_view>
 
+// since cib_data_fmt is shared with the board
+// we need this macro to avoid clashes with the namespaces
+// not the cleanest solution, but it works and avoids maintaining two separate versions of the same struct
 #define CIB_DUNEDAQ 1
 // we may need this to help parse the data format arriving from the CIB
 #include <cib_data_fmt.h>
-
+#include <cib_utilities.h>
 /**
  * @brief Name used by TRACE TLOG calls from this source file
  */
 #define TRACE_NAME "CIBModule" // NOLINT
 #define TLVL_ENTER_EXIT_METHODS 10
 #define TLVL_CIB_INFO 5
-#define TLVL_CIB_DEBUG 15
-#define TLVL_CIB_DEBUG 15
+#define TLVL_CIB_DEBUG 10
+#define TLVL_CIB_TRACE 15
 
+// as it happens, the frame structure remains unchanged since  NP04
 constexpr uint16_t CIB_HSI_FRAME_VERSION = 0x1; // NOLINT
 namespace dunedaq::cibmodules {
-
-  // some helper functions outside of the class
-  // taken from the cib_data_utils
-  namespace util {
-
-    uint32_t bitmask(uint32_t highbit, uint32_t lowbit)
-    {
-      // sort the bit order or this fails miserably
-      if (highbit < lowbit)
-      {
-        uint32_t tmp = lowbit;
-        lowbit = highbit;
-        highbit = tmp;
-      }
-
-      uint32_t i = ~0U;
-      return ~(i << highbit << 1) & (i << lowbit);
-    }
-    
-    // converts a masked unsigned value into a signed
-    // the mask is always assumed to start at 0, so the value has to be shifted right until the lsb aligns with 0
-    int32_t cast_to_signed(const uint32_t reg, const uint32_t mask)
-    {
-      // first find the msb in the mask. That will be the signed bit
-      uint32_t msb = 0;
-      int32_t res = 0;
-      for (size_t bit = 31; bit > 0; bit--)
-      {
-        if ((1U << bit) & mask)
-        {
-          msb = bit;
-          break;
-        }
-      }
-      // spdlog::trace("MSB of the mask is {0}",msb);
-      //  check the msb of the register. That is the sign bit
-      if ((1U << msb) & reg)
-      {
-        // spdlog::trace("MSB of the mask is {0}",msb);
-
-        res = bitmask(31, msb + 1); // set all bits to 1 above the mask
-        res = res | (reg & mask);
-        // it is a negative value. Set the msb in the result
-      }
-      else
-      {
-        // it is a positive value. No need to set the sign bit, but still need to
-        // apply the mask or we're carrying out the other bits that may be outside the mask
-        res = (reg & mask);
-      }
-      return res;
-    }
-
-    int32_t get_m1(dunedaq::cib::daq::iols_trigger_t &t)
-    {
-      return cast_to_signed(t.pos_m1, t.bitmask_m1);
-    }
-
-    int32_t get_m2(dunedaq::cib::daq::iols_trigger_t &t)
-    {
-      uint32_t m2_lsb = t.pos_m2_lsb;
-      uint32_t m2_msb = t.pos_m2_msb;
-      uint32_t m2 = (m2_msb << 15) | t.pos_m2_lsb;
-      // the bitmask is the same
-      return cast_to_signed(m2, t.bitmask_m2);
-    }
-
-    int32_t get_m3(dunedaq::cib::daq::iols_trigger_t &t)
-    {
-      return cast_to_signed(t.pos_m3, t.bitmask_m3);
-    }
-  } // namespace util
-
-
-
 
   CIBModule::CIBModule(const std::string& name)
               : hsilibs::HSIEventSender(name)
@@ -236,7 +165,7 @@ namespace dunedaq::cibmodules {
 
     // identify the trigger bit that this receiver is assigned to
     // We need this to construct the HSI frame, right?
-    if (!parse_hex(trigger_conf->get_trigger_bit(), m_trigger_bit))
+    if (!dunedaq::cibmodules::util::parse_hex(trigger_conf->get_trigger_bit(), m_trigger_bit))
     {
       std::ostringstream msg("");
       msg << get_name() << ": Unable to parse trigger bit hex string : " << trigger_conf->get_trigger_bit();
@@ -244,30 +173,23 @@ namespace dunedaq::cibmodules {
     }
     else
     {
-      TLOG_DEBUG(TLVL_CIB_INFO) << get_name() << ": Parsed trigger bit hex string "
+      TLOG_DEBUG(TLVL_CIB_DEBUG) << get_name() << ": Parsed trigger bit hex string "
                                 << trigger_conf->get_trigger_bit() << " to 0x"
                                 << std::hex << m_trigger_bit << std::dec
                                 << "[" << trigger_conf->get_trigger_id() << "]";
     }
-    
-    // m_trigger_bit = trigger_conf->get_trigger_bit();
-    // m_module_instance = conf->get_instance();
-    // TLOG_DEBUG(TLVL_CIB_INFO) << get_name() << ": Instance assigned to trigger bit " << m_trigger_bit
-    //     << " ( 0x" << std::hex << m_trigger_bit << std::dec << ")";
     
     // init monitoring variables
     m_num_control_messages_sent = 0;
     m_num_control_responses_received = 0;
 
     // figure out the identifier of the CIB
-    // this is set in the configuration, right?
-    // auto board = m_module->get_board();
+    // this is set in the configuration
     auto geo_id = board->get_geo_id();
     m_det = geo_id->get_detector_id();
     m_crate = geo_id->get_crate_id();
     m_slot = geo_id->get_slot_id();
 
-    // const auto& misc = board->get_misc();
     auto session = m_cfg->get_session();
 
     // init trigger counters
@@ -413,7 +335,7 @@ namespace dunedaq::cibmodules {
       TLOG_DEBUG(TLVL_CIB_INFO) << get_name() << ": Sending start of run command with run number " << m_run_number.load();
       m_thread_.start_working_thread();
 
-      // NFB: There is a potential race condition here: the socket in the working thread
+      // NFB: There was a potential race condition here: the socket in the working thread
       // needs to be in place before the CIB receives order to send data, or we risk having a connection
       // failure, if for some reason the CIB attempts to connect before the working thread is ready to receive.
       if (m_calibration_stream_enable)
@@ -505,9 +427,6 @@ namespace dunedaq::cibmodules {
     //connect to socket
     // should we keep everything local or under the class?
     boost::system::error_code ec;
-    //boost::asio::ip::tcp::endpoint( boost::asio::ip::tcp::v4(),m_receiver_port )
-
-    // unsigned short port = m_receiver_port;
 
     // check that this port is still available
     if (check_port_in_use(m_receiver_port))
@@ -553,10 +472,6 @@ namespace dunedaq::cibmodules {
       {
         break ;
       }
-//      else
-//      {
-//        TLOG_DEBUG(TLVL_DEBUG_MEDIUM) << "Waiting for a bit longer";
-//      }
     }
 
     TLOG() << get_name() <<  ": Connection received: start reading" << std::endl;
@@ -569,7 +484,7 @@ namespace dunedaq::cibmodules {
      */
 
     dunedaq::cib::daq::iols_tcp_packet_t tcp_packet;
-    TLOG_DEBUG(TLVL_CIB_DEBUG) << "Checking expected sizes: "
+    TLOG_DEBUG(TLVL_CIB_TRACE) << "Checking expected sizes: "
                                << " sizeof(iols_tcp_packet_t)=" << sizeof(dunedaq::cib::daq::iols_tcp_packet_t)
                                << " sizeof(iols_trigger_t)=" << sizeof(dunedaq::cib::daq::iols_trigger_t)
                                << " sizeof(tcp_header_t)=" << sizeof(dunedaq::cib::daq::tcp_header_t)
@@ -641,7 +556,8 @@ namespace dunedaq::cibmodules {
       update_buffer_counts(n_words);
 
       // temporarily print the trigger
-      // TLOG_DEBUG(TLVL_CIB_DEBUG) << "TRIGGER : ts " << tcp_packet.word.timestamp
+      // I leave this here for debugging purposes, but don´t even keep it in the logs, since it is too verbose. We can always add it back if we need to debug something
+      // TLOG_DEBUG(TLVL_CIB_TRACE) << "TRIGGER : ts " << tcp_packet.word.timestamp
       //                            << " pos_m1 " << util::get_m1(tcp_packet.word)
       //                            << " pos_m2 " << util::get_m2(tcp_packet.word)
       //                            << " pos_m3 " << util::get_m3(tcp_packet.word);
@@ -676,9 +592,9 @@ namespace dunedaq::cibmodules {
 
       // we shall use these 2 sets of 32 bits to define the periscope position
       // pos_m3 == linear stage
-      hsi_struct[3] = tcp_packet.word.pos_m3; // lower 32b 0
+      hsi_struct[3] = dunedaq::cibmodules::util::get_m3(tcp_packet.word); // lower 32b 0
       // pos_m3 == RNN600
-      hsi_struct[4] = tcp_packet.word.pos_m2_msb << 15 | tcp_packet.word.pos_m2_lsb; // upper 32b
+      hsi_struct[4] = dunedaq::cibmodules::util::get_m2(tcp_packet.word); // upper 32b
       /**
        * A note about the 5th entry
        * The trigger bit is actually mapped into a single bit, that is then remapped back
@@ -687,7 +603,8 @@ namespace dunedaq::cibmodules {
       hsi_struct[5] = m_trigger_bit;            // trigger_map;
       hsi_struct[6] = m_num_run_triggers_received.load();    // m_generated_counter;
 
-      // TLOG_DEBUG(TLVL_CIB_DEBUG) << "CIB HSI Frame: "
+      // Same thing here. If something really bad happens, this code can be very useful, but we do not want to print this in the logs, unless strictly necessary, since it is too verbose. 
+      // TLOG_DEBUG(TLVL_CIB_TRACE) << "CIB HSI Frame: "
       //                            << "0x" << std::hex << hsi_struct[0]
       //                            << ", 0x" << hsi_struct[1]
       //                            << ", 0x" << hsi_struct[2]
@@ -708,8 +625,10 @@ namespace dunedaq::cibmodules {
 
       send_raw_hsi_data(hsi_struct, m_cib_hsi_data_sender.get());
 
-      // TODO Nuno Barros Apr-02-2024 : properly fill device id
-      // still need to figure this one out.
+      // TODO Nuno Barros Apr-02-2024 : properly fill device id when someone explains me
+      // how to get it
+      // in fact, this may disappear in the future where the CIB is decoupled from 
+      // the HSI infrastructure
       dfmessages::HSIEvent event(m_det,
                                  m_trigger_bit,
                                  tcp_packet.word.timestamp,
@@ -824,8 +743,6 @@ namespace dunedaq::cibmodules {
       return ;
     }
     m_last_calibration_file_update = std::chrono::steady_clock::now();
-    // _calibration_file.setf ( std::ios::hex, std::ios::basefield );
-    // _calibration_file.unsetf ( std::ios::showbase );
     TLOG_DEBUG(TLVL_CIB_INFO) << get_name() << ": New Calibration Stream file: " << global_name << std::endl ;
   }
 
@@ -870,8 +787,6 @@ namespace dunedaq::cibmodules {
     TLOG_DEBUG(TLVL_CIB_INFO) << get_name() << ": Sending config" << std::endl;
 
     // structure the message to have a common management structure
-    //json receiver = doc.at("ctb").at("sockets").at("receiver");
-
     nlohmann::json conf;
     conf["command"] = "config";
     conf["config"] = nlohmann::json::parse(config);
@@ -1008,24 +923,6 @@ namespace dunedaq::cibmodules {
     return ec == error::address_in_use;
 
   }
-
-
-  bool CIBModule::parse_hex(std::string_view s, std::uint32_t &out)
-  {
-    // Optional 0x / 0X prefix
-    if (s.size() >= 2 && s[0] == '0' && (s[1] == 'x' || s[1] == 'X'))
-      s.remove_prefix(2);
-
-    // Empty after stripping?
-    if (s.empty())
-      return false;
-
-    auto [ptr, ec] = std::from_chars(s.data(), s.data() + s.size(), out, 16);
-
-    // ec=={} means parse OK; ptr at end means no trailing garbage
-    return ec == std::errc{} && ptr == s.data() + s.size();
-  }
-
 } // namespace dunedaq::cibmodules
 
 DEFINE_DUNE_DAQ_MODULE(dunedaq::cibmodules::CIBModule)
